ref: inner train loop (intermediate step) 21/n
